requirement.txt
pytorch 1.4.0
torchvision 0.5.0
geopandas
PIL
shapely
numpy
opencv